Sitting HR Lady Implemented
The HR lady's sitting animation works. It's boring, but it gets the job done.
Few notes: Character looks a bit funky, guessing that's because I didn't do the whole import textures thing with the Character Creator addon, so that still needs to be done. IF the coloring is an issue with the model itself, I currently can't open Character Creator. It's giving me this weird message every time I open it up, so I'm not sure what to do about that.